start moving common functions for tree sequence examples to separate sources
