+++ examples/wfts.cc
@@ -38,146 +38,12 @@
 #include "simplify_tables.hpp"
 #include "evolve_generation_ts.hpp"
 #include "calculate_fitnesses.hpp"
+#include "tree_sequence_examples_common.hpp"
 
 namespace po = boost::program_options;
 using poptype = fwdpp::slocuspop<fwdpp::popgenmut>;
-using GSLrng = fwdpp::GSLrng_t<fwdpp::GSL_RNG_MT19937>;
+using GSLrng = fwdpp::GSLrng_mt;
 
-struct diploid_metadata
-{
-    std::size_t individual;
-    double time, fitness;
-    fwdpp::ts::TS_NODE_INT n1, n2;
-    diploid_metadata(std::size_t i, double t, double w,
-                     fwdpp::ts::TS_NODE_INT a, fwdpp::ts::TS_NODE_INT b)
-        : individual(i), time(t), fitness(w), n1(a), n2(b)
-    {
-    }
-};
-
-void
-expensive_leaf_test(const fwdpp::ts::table_collection &tables,
-                    const std::vector<fwdpp::ts::TS_NODE_INT> &sample_list)
-{
-    fwdpp::ts::tree_visitor mti(tables, sample_list);
-    while (mti(std::true_type(), std::true_type()))
-        {
-            auto &tree = mti.tree();
-            for (auto i : sample_list)
-                {
-                    auto p = i;
-                    while (p != -1)
-                        {
-                            auto l = tree.left_sample[p];
-                            auto ogl = l;
-                            if (l != -1)
-                                {
-                                    auto r = tree.right_sample[p];
-                                    int ns = 0;
-                                    while (true)
-                                        {
-                                            ++ns;
-                                            if (l == r)
-                                                {
-                                                    break;
-                                                }
-                                            l = tree.next_sample[l];
-                                            if (l == ogl)
-                                                {
-                                                    throw std::runtime_error(
-                                                        "loopback error");
-                                                }
-                                        }
-                                    if (ns != tree.leaf_counts[p])
-                                        {
-                                            throw std::runtime_error(
-                                                "bad sample interval");
-                                        }
-                                }
-                            p = tree.parents[p];
-                        }
-                }
-        }
-}
-
-template <typename mcont_t>
-void
-matrix_runtime_test(const fwdpp::ts::table_collection &tables,
-                    const std::vector<fwdpp::ts::TS_NODE_INT> &samples,
-                    const mcont_t &mutations,
-                    const std::vector<fwdpp::uint_t> &mcounts)
-{
-    auto dm = fwdpp::ts::generate_data_matrix(tables, samples, mutations, true,
-                                              true);
-    auto rs = fwdpp::row_sums(dm);
-    for (std::size_t i = 0; i < rs.first.size(); ++i)
-        {
-            if (rs.first[i] != mcounts[dm.neutral_keys[i]])
-                {
-                    throw std::runtime_error("bad neutral mutation count");
-                }
-        }
-    for (std::size_t i = 0; i < rs.second.size(); ++i)
-        {
-            if (rs.second[i] != mcounts[dm.selected_keys[i]])
-                {
-                    throw std::runtime_error("bad selected mutation count");
-                }
-        }
-}
-
-void
-test_serialization(const fwdpp::ts::table_collection &tables,
-                   const std::string &filename)
-{
-    std::ofstream o(filename.c_str());
-    fwdpp::ts::io::serialize_tables(o, tables);
-    o.close();
-    std::ifstream i(filename.c_str());
-    auto tables2 = fwdpp::ts::io::deserialize_tables(i);
-
-    if (tables.genome_length() != tables2.genome_length())
-        {
-            throw std::runtime_error("genome_length does not match");
-        }
-    if (tables.edge_offset != tables2.edge_offset)
-        {
-            throw std::runtime_error("edge_offset does not match");
-        }
-
-    if (tables.edge_table != tables2.edge_table)
-        {
-            throw std::runtime_error("edge tables do not match");
-        }
-
-    if (tables.node_table != tables2.node_table)
-        {
-            throw std::runtime_error("node tables do not match");
-        }
-
-    if (tables.mutation_table != tables2.mutation_table)
-        {
-            throw std::runtime_error("mutation tables do not match");
-        }
-    if (tables != tables2)
-        {
-            throw std::runtime_error("tables failed equality check");
-        }
-}
-
-template <typename rng>
-std::function<double()>
-make_dfe(const fwdpp::uint_t N, const rng &r, const double mean,
-         const double shape, const double scoeff)
-{
-    if (std::isfinite(scoeff))
-        {
-            return [scoeff]() { return scoeff; };
-        }
-    fwdpp::extensions::gamma dfe(mean, shape);
-    return
-        [&r, dfe, N]() { return dfe(r.get()) / static_cast<double>(2 * N); };
-}
 
 int
 main(int argc, char **argv)
